// src/app/models/ambulancia.model.ts
export class AmbulanciaModel {
    uid?: string;
    id?: string;
    tipo: string;
    matricula: string;
    capacidad: string;
    marca: string;
    modelo: string;
    equipamiento: string;
    horario_servicio: string;
    conductor: string;
    numero_de_licencia: string;
    celular_de_contacto: string;
    responsable: string;
    key$?: string;
    lat: number;
    lng: number;
}